[_toplevel.trace] Make trace functions better follow their original behaviour
The documentation for this is scatter across multiple places and isn't very clear. I think this is better but could be misinterpretting it.

+++ as3lib/_toplevel/trace.py
@@ -1,30 +1,28 @@
 from as3lib import as3state
 
 
+def _traceFileOutput(output):
+   if as3state.TraceOutputFileName.exists() and as3state.TraceOutputFileName.is_file():
+      with open(as3state.TraceOutputFileName, 'a') as f:
+         f.write(output + '\n')
+   else:
+      with open(as3state.TraceOutputFileName, 'w') as f:
+         f.write(output + '\n')
+
 def trace(*args):
    if as3state.as3DebugEnable:
       output = ' '.join((str(i) for i in args))
+      print(output)
       if as3state.TraceOutputFileEnable:
-         if as3state.TraceOutputFileName.exists():
-            if as3state.TraceOutputFileName.is_file():
-               with open(as3state.TraceOutputFileName, 'a') as f:
-                  f.write(output + '\n')
-            else:
-               print(output)
-         else:
-            with open(as3state.TraceOutputFileName, 'w') as f:
-               f.write(output + '\n')
-      else:
-         print(output)
+         _traceFileOutput(output)
 
 def errorTrace(*args):
-   '''
-   This is a wrapper around trace that increments as3state.CurrentWarnings 
-   '''
-   if as3state.ErrorReportingEnable and not as3state.MaxWarningsReached:
+   output = ' '.join(str(i) for i in args)
+   print(output)
+   if as3state.as3DebugEnable and as3state.ErrorReportingEnable and not as3state.MaxWarningsReached:
       if as3state.CurrentWarnings < as3state.MaxWarnings or as3state.MaxWarnings == 0:
-         trace(*args)
          as3state.CurrentWarnings += 1
       else:
-         trace('Maximum number of errors has been reached. All further errors will be suppressed.')
+         output = 'Maximum number of errors has been reached. All further errors will be suppressed.'
          as3state.MaxWarningsReached = True
+      _traceFileOutput(output)
